\name{BalanceCheck}
\alias{BalanceCheck}
\title{Balance Check for Multiple Covariates in Matched Observational Studies}
\description{This package includes two tests for assessing whether multiple covariates in a treatment group and a matched control group are balanced in observational studies.
}
\author{
Hao Chen and Dylan Small

Maintainer: Hao Chen (hxchen@ucdavis.edu)
}
\references{
Chen, H. and Small, D. (2016) New multivariate tests for assessing covariate balance in matched observational studies.
}

\seealso{
  \code{\link{CrossMST}}, \code{\link{CrossNN}}
}

