\name{smokingDist}
\alias{smokingDist}
\title{The distance matrix of the smoking data}
\description{This is a 300 by 300 distance matrix with the indices of the treated subjects 1:150 and the indices of the matched controls 151:300.  These treated subjects and matched controls are a subset of the original dataset used in the reference paper with 679 treated subjects and 679 matched controls.}

